Struct spec_lang::code_writer::CodeWriter

pub struct CodeWriter(_);

A helper to emit code. Supports indentation and maintains source to target location information.

Methods

impl CodeWriter

pub fn new(loc: Loc) -> CodeWriter

Creates new code writer, with the given default location.

pub fn create_label(&self) -> CodeWriterLabel

Creates a label at which code can be inserted later.

pub fn insert_at_label(&self, label: CodeWriterLabel, s: &str)

Inserts code at the previously created label.

pub fn process_result<T, F: FnMut(&str) -> T>(&self, f: F) -> T

Calls a function to process the code written so far. This is embedded into a function so we ensure correct scoping of borrowed RefCell content.

pub fn set_location(&self, loc: &Loc)

Sets the current location. This location will be associated with all subsequently written code so we can map back from the generated code to this location. If current loc is already the passed one, nothing will be updated, so it is ok to call this method repeatedly with the same value.

pub fn get_source_location(&self, output_index: ByteIndex) -> Option<Loc>

Given a byte index in the written output, return the best approximation of the source which generated this output.

pub fn get_output_byte_index(
&self,
line: LineIndex,
column: ColumnIndex
) -> Option<ByteIndex>

Given line/column location, determine ByteIndex of that location.

pub fn indent(&self)

Indents any subsequently written output. The current line of output and any subsequent ones will be indented. Note this works after the last output was \n but the line is still empty.

pub fn unindent(&self)

Undo previously done indentation.

pub fn emit(&self, s: &str)

Emit a string. The string will be broken down into lines to apply current indentation.

pub fn emit_line(&self, s: &str)

Emits a string and then terminates the line.