Module spec_lang::env

Provides an environment -- global state -- for translation, including helper functions to interpret metadata about the translation target.

Structs

FieldData
Field Environment

FieldEnv

FieldId
Identifier for a field of a structure, relative to struct.

FunId
Identifier for a Move function, relative to module.

FunctionData

FunctionEnv

GlobalEnv
Global Environment

Loc
Locations

ModuleData
Module Environment

ModuleEnv
Represents a module environment.

ModuleId
Identifier for a module.

NodeId
Identifier for a node in the AST, relative to a module. This is used to associate attributes with the node, like source location and type.

Parameter
Represents a parameter.

SchemaId
Identifier for a schema.

SpecFunId
Identifier for a specification function, relative to module.

SpecVarId
Identifier for a specification variable, relative to module.

StructData
Struct Environment

StructEnv

StructId
Identifier for a structure/resource, relative to module.

TypeParameter
Function Environment

Enums

TypeConstraint

Constants

SCRIPT_AST_FUN_NAME

SCRIPT_BYTECODE_FUN_NAME
Names used in the bytecode/AST to represent the main function of a script

SCRIPT_MODULE_NAME
Constants

Type Definitions

MoveIrLoc
Alias for the Loc variant of MoveIR. This uses a &static str instead of FileId for the file name.