Struct spec_lang::env::GlobalEnv

pub struct GlobalEnv { /* fields omitted */ }

Global Environment

Global environment for a set of modules.

Methods

impl GlobalEnv

pub fn new() -> Self

Creates a new environment.

pub fn symbol_pool(&self) -> &SymbolPool

Returns a reference to the symbol pool owned by this environment.

pub fn add_source(
&mut self,
file_name: &str,
source: &str,
is_dep: bool
) -> FileId

Adds a source to this environment, returning a FileId for it.

pub fn add_diag(&self, diag: Diagnostic)

Adds diagnostic to the environment.

pub fn error_with_notes(&self, loc: &Loc, msg: &str, notes: Vec<String>)

Adds an error to this environment, with notes.

pub fn error(&self, loc: &Loc, msg: &str)

Adds an error to this environment, without notes.

pub fn unknown_loc(&self) -> Loc

Returns the unknown location.

pub fn unknown_move_ir_loc(&self) -> MoveIrLoc

Returns a Move IR version of the unknown location which is guaranteed to map to the regular unknown location via to_loc.

pub fn internal_loc(&self) -> Loc

Returns the internal location.

pub fn to_loc(&self, loc: &MoveIrLoc) -> Loc

Converts a Loc as used by the move-lang compiler to the one we are using here. TODO: move-lang should use FileId as well so we don't need this here. There is already a todo in their code to remove the current use of &'static str for file names in Loc.

pub fn file_id_to_idx(&self, file_id: FileId) -> u16

Maps a FileId to an index which can be mapped back to a FileId.

pub fn file_idx_to_id(&self, file_idx: u16) -> FileId

Maps a an index which was obtained by file_id_to_idx back to a FileId.

pub fn get_position(&self, loc: Loc) -> Option<(String, Location)>

Returns file name and line/column position for a location, if available.

pub fn get_file_count(&self) -> usize

pub fn has_errors(&self) -> bool

Returns true if diagnostics have error severity or worse.

pub fn report_errors<W: WriteColor>(&self, writer: &mut W)

Writes accumulated diagnostics to writer.

pub fn add(
&mut self,
loc: Loc,
module: CompiledModule,
source_map: SourceMap<MoveIrLoc>,
struct_data: BTreeMap<StructId, StructData>,
function_data: BTreeMap<FunId, FunctionData>,
spec_vars: Vec<SpecVarDecl>,
spec_funs: Vec<SpecFunDecl>,
module_spec: Spec,
loc_map: BTreeMap<NodeId, Loc>,
type_map: BTreeMap<NodeId, Type>,
instantiation_map: BTreeMap<NodeId, Vec<Type>>
)

Adds a new module to the environment. StructData and FunctionData need to be provided in definition index order. See create_function_data and create_struct_data for how to create them.

pub fn create_function_data(
&self,
module: &CompiledModule,
def_idx: FunctionDefinitionIndex,
name: Symbol,
loc: Loc,
arg_names: Vec<Symbol>,
type_arg_names: Vec<Symbol>,
spec: Spec
) -> FunctionData

Creates data for a function, adding any information not contained in bytecode. This is a helper for adding a new module to the environment.

pub fn create_struct_data(
&self,
module: &CompiledModule,
def_idx: StructDefinitionIndex,
name: Symbol,
loc: Loc,
spec: Spec
) -> StructData

Creates data for a struct. Currently all information is contained in the byte code. This is a helper for adding a new module to the environment.

pub fn find_module(&self, name: &ModuleName) -> Option<ModuleEnv>

Finds a module by name and returns an environment for it.

pub fn find_module_by_name(&self, simple_name: Symbol) -> Option<ModuleEnv>

Finds a module by simple name and returns an environment for it. TODO: we may need to disallow this to support modules of the same simple name but with different addresses in one verification session.

pub fn get_enclosing_module(&self, loc: Loc) -> Option<ModuleEnv>

Return the module enclosing this location.

pub fn get_enclosing_function(&self, loc: Loc) -> Option<FunctionEnv>

Returns the function enclosing this location.

pub fn get_enclosing_struct(&self, loc: Loc) -> Option<StructEnv>

Returns the struct enclosing this location.

pub fn get_module_count(&self) -> usize

pub fn get_module(&self, id: ModuleId) -> ModuleEnv

Gets a module by id.

pub fn get_modules(&self) -> impl Iterator<Item = ModuleEnv>

Returns an iterator for all modules in the environment.

pub fn get_bytecode_modules(&self) -> impl Iterator<Item = &CompiledModule>

Returns an iterator for all bytecode modules in the environment.

pub fn get_all_structs_with_conditions(&self) -> Vec<Type>

Returns all structs in all modules which carry invariants.