Struct spec_lang::env::ModuleData

pub struct ModuleData {
    pub name: ModuleName,
    pub id: ModuleId,
    pub module: CompiledModule,
    pub struct_data: BTreeMap<StructId, StructData>,
    pub struct_idx_to_id: BTreeMap<StructDefinitionIndex, StructId>,
    pub function_data: BTreeMap<FunId, FunctionData>,
    pub function_idx_to_id: BTreeMap<FunctionDefinitionIndex, FunId>,
    pub spec_vars: BTreeMap<SpecVarId, SpecVarDecl>,
    pub spec_funs: BTreeMap<SpecFunId, SpecFunDecl>,
    pub module_spec: Spec,
    pub source_map: SourceMap<MoveIrLoc>,
    pub loc: Loc,
    pub loc_map: BTreeMap<NodeId, Loc>,
    pub type_map: BTreeMap<NodeId, Type>,
    pub instantiation_map: BTreeMap<NodeId, Vec<Type>>,
    pub spec_block_infos: Vec<SpecBlockInfo>,
}

Module Environment

Represents data for a module.

Fields
name: ModuleName

Module name.

id: ModuleId

Id of this module in the global env.

module: CompiledModule

Module byte code.

struct_data: BTreeMap<StructId, StructData>

Struct data.

struct_idx_to_id: BTreeMap<StructDefinitionIndex, StructId>

Mapping from struct definition index to id in above map.

function_data: BTreeMap<FunId, FunctionData>

Function data.

function_idx_to_id: BTreeMap<FunctionDefinitionIndex, FunId>

Mapping from function definition index to id in above map.

spec_vars: BTreeMap<SpecVarId, SpecVarDecl>

Specification variables, in SpecVarId order.

spec_funs: BTreeMap<SpecFunId, SpecFunDecl>

Specification functions, in SpecFunId order.

module_spec: Spec

Module level specification.

source_map: SourceMap<MoveIrLoc>

Module source location information.

loc: Loc

The location of this module.

loc_map: BTreeMap<NodeId, Loc>

A map from node id to associated location.

type_map: BTreeMap<NodeId, Type>

A map from node id to associated type.

instantiation_map: BTreeMap<NodeId, Vec<Type>>

A map from node id to associated instantiation of type parameters.

spec_block_infos: Vec<SpecBlockInfo>

A list of spec block infos, for documentation generation.