Struct spec_lang::env::ModuleEnv

pub struct ModuleEnv<'env> {
    pub env: &'env GlobalEnv,
    // some fields omitted
}

Represents a module environment.

Fields
env: &'env GlobalEnv

Reference to the outer env.

Methods

impl<'env> ModuleEnv<'env>

pub fn get_id(&self) -> ModuleId

Returns the id of this module in the global env.

pub fn get_name(&'env self) -> &'env ModuleName

Returns the name of this module.

pub fn is_script_module(&self) -> bool

Returns true if this is a module representing a script.

pub fn is_dependency(&self) -> bool

Returns true of this module is from a dependency, i.e. not the target of verification.

pub fn get_source_path(&self) -> &OsStr

Returns the path to source file of this module.

pub fn get_doc(&self) -> &str

Returns documentation associated with this module.

pub fn get_spec_block_infos(&self) -> &[SpecBlockInfo]

Returns spec block documentation infos.

pub fn symbol_pool(&self) -> &SymbolPool

Shortcut for accessing the symbol pool.

pub fn get_verified_module(&'env self) -> &'env CompiledModule

Gets the underlying bytecode module.

pub fn find_function(&self, name: Symbol) -> Option<FunctionEnv<'env>>

Gets a FunctionEnv in this module by name.

pub fn get_function(&'env self, id: FunId) -> FunctionEnv<'env>

Gets a FunctionEnv by id.

pub fn into_function(self, id: FunId) -> FunctionEnv<'env>

Gets a FunctionEnv by id.

pub fn get_function_count(&self) -> usize

Gets the number of functions in this module.

pub fn get_functions(&'env self) -> impl Iterator<Item = FunctionEnv<'env>>

Returns iterator over FunctionEnvs in this module.

pub fn into_functions(self) -> impl Iterator<Item = FunctionEnv<'env>>

Returns iterator over FunctionEnvs in this module.

pub fn get_called_function(&self, idx: FunctionHandleIndex) -> FunctionEnv

Get FunctionEnv for a function used in this module, via the FunctionHandleIndex. The returned function might be from this or another module.

pub fn find_struct(&self, name: Symbol) -> Option<StructEnv>

Gets a StructEnv in this module by name.

pub fn get_struct_id(&self, idx: StructDefinitionIndex) -> StructId

Gets the struct id from a definition index which must be valid for this environment.

pub fn get_struct(&self, id: StructId) -> StructEnv

Gets a StructEnv by id.

pub fn get_struct_by_def_idx(&self, idx: StructDefinitionIndex) -> StructEnv

pub fn get_function_id(&self, idx: FunctionDefinitionIndex) -> FunId

Gets the function id from a definition index which must be valid for this environment.

pub fn into_struct(self, id: StructId) -> StructEnv<'env>

Gets a StructEnv by id, consuming this module env.

pub fn get_struct_count(&self) -> usize

Gets the number of structs in this module.

pub fn get_structs(&'env self) -> impl Iterator<Item = StructEnv<'env>>

Returns iterator over structs in this module.

pub fn into_structs(self) -> impl Iterator<Item = StructEnv<'env>>

Returns iterator over structs in this module.

pub fn globalize_signature(&self, sig: &SignatureToken) -> Type

Globalizes a signature local to this module.

pub fn globalize_signatures(&self, sigs: &[SignatureToken]) -> Vec<Type>

Globalizes a list of signatures.

pub fn get_type_actuals(&self, idx: Option<SignatureIndex>) -> Vec<Type>

Gets a list of type actuals associated with the index in the bytecode.

pub fn get_constant(&self, idx: ConstantPoolIndex) -> &VMConstant

Retrieve a constant from the pool

pub fn get_constant_value(&self, constant: &VMConstant) -> MoveValue

Converts a constant to the specified type. The type must correspond to the expected cannonical representation as defined in move_core_types::values

pub fn get_address_identifier(&self, idx: AddressIdentifierIndex) -> BigUint

Retrieve an address identifier from the pool

pub fn addr_to_big_uint(addr: &AccountAddress) -> BigUint

Converts an address identifier to a number representing the address.

pub fn get_spec_vars(
&'env self
) -> impl Iterator<Item = (&'env SpecVarId, &'env SpecVarDecl)>

Returns specification variables of this module.

pub fn get_spec_var(&self, id: SpecVarId) -> &SpecVarDecl

Gets spec var by id.

pub fn find_spec_var(&self, name: Symbol) -> Option<&SpecVarDecl>

Find spec var by name.

pub fn get_spec_funs(
&'env self
) -> impl Iterator<Item = (&'env SpecFunId, &'env SpecFunDecl)>

Returns specification functions of this module.

pub fn get_spec_fun(&self, id: SpecFunId) -> &SpecFunDecl

Gets spec fun by id.

pub fn get_spec(&self) -> &Spec

Gets module specification.

pub fn get_spec_funs_of_name(
&self,
name: Symbol
) -> impl Iterator<Item = (&'env SpecFunId, &'env SpecFunDecl)>

Get all spec fun overloads with the given name.

pub fn get_node_loc(&self, node_id: NodeId) -> Loc

Gets the location of the given node.

pub fn get_node_type(&self, node_id: NodeId) -> Type

Gets the type of the given node.

pub fn get_node_instantiation(&self, node_id: NodeId) -> Vec<Type>

Gets the type parameter instantiation associated with the given node.