Enum spec_lang::ty::Type

pub enum Type {
    Primitive(PrimitiveType),
    Tuple(Vec<Type>),
    Vector(Box<Type>),
    Struct(ModuleId, StructId, Vec<Type>),
    TypeParameter(u16),
    Reference(bool, Box<Type>),
    Fun(Vec<Type>, Box<Type>),
    TypeDomain(Box<Type>),
    TypeLocal(Symbol),
    Error,
    Var(u16),
}

Represents a type.

Variants

Primitive(PrimitiveType)

Tuple(Vec<Type>)

Vector(Box<Type>)

Struct(ModuleId, StructId, Vec<Type>)

TypeParameter(u16)

Reference(bool, Box<Type>)

Fun(Vec<Type>, Box<Type>)

TypeDomain(Box<Type>)

TypeLocal(Symbol)

Error

Var(u16)

Methods

impl Type

pub fn new_prim(p: PrimitiveType) -> Type

pub fn is_reference(&self) -> bool

Determines whether this is a reference.

pub fn is_mutable_reference(&self) -> bool

Determines whether this is a mutable reference.

pub fn is_immutable_reference(&self) -> bool

Determines whether this is an immutable reference.

pub fn is_number(&self) -> bool

Returns true if this is any number type.

pub fn replace_struct_instantiation(&self, inst: &[Type]) -> Type

If this is a struct type, replace the type instantiation.

pub fn get_struct<'env>(
&'env self,
env: &'env GlobalEnv
) -> Option<(StructEnv<'env>, &'env [Type])>

If this is a struct type, return the associated struct env and type parameters.

pub fn instantiate(&self, params: &[Type]) -> Type

Instantiates type parameters in this type.

pub fn contains<P>(&self, p: &P) -> bool where
P: Fn(&Type) -> bool,

Checks whether this type contains a type for which the predicate is true.

pub fn is_incomplete(&self) -> bool

Returns true if this type is incomplete, i.e. contains any type variables.

impl Type

pub fn display<'a>(
&'a self,
context: &'a TypeDisplayContext<'a>
) -> TypeDisplay<'a>